+++ backend/canisters/community/api/src/updates/c2c_join_community.rs
@@ -1,6 +1,6 @@
 use candid::{CandidType, Principal};
 use serde::{Deserialize, Serialize};
-use types::{CommunityCanisterCommunitySummary, GateCheckFailedReason, UserId};
+use types::{CommunityCanisterCommunitySummary, GateCheckFailedReason, TimestampMillis, UserId};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
@@ -9,6 +9,7 @@ pub struct Args {
     pub invite_code: Option<u64>,
     pub is_platform_moderator: bool,
     pub is_bot: bool,
+    pub diamond_membership_expires_at: Option<TimestampMillis>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]