+++ backend/canisters/local_user_index/impl/src/queries/mod.rs
@@ -1,5 +1,6 @@
 pub mod c2c_can_push_notifications;
 pub mod c2c_lookup_user;
+pub mod c2c_lookup_users;
 pub mod c2c_user_principals;
 pub mod chat_events;
 pub mod group_and_community_summary_updates;